backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# API & Data validation
pydantic==2.9.0
pydantic-settings==2.6.0

# AI & Chatbot (Choose one or install all for flexibility)
openai==1.54.0                    # For OpenAI GPT models
anthropic==0.40.0                 # For Claude models
google-generativeai==0.8.3        # For Google Gemini models
langchain==0.3.7
langchain-openai==0.2.8

# Payment
stripe==11.2.0

# Utilities
httpx==0.27.2
aiofiles==24.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
